fix (addon): #663 rebuild the places cache after 30 minutes
