+++ lib/ActivityStreams.js
@@ -41,6 +41,7 @@ const DEFAULT_OPTIONS = {
   onAddWorker: null,
   onRemoveWorker: null,
   previewCacheTimeout: 21600000, // every 6 hours, rebuild/repopulate the cache
+  placesCacheTimeout: 1800000, // every 30 minutes, rebuild/repopulate the cache
 };
 
 function ActivityStreams(options = {}) {
@@ -314,6 +315,16 @@ ActivityStreams.prototype = {
         this._populatingCache.places = false;
         Services.obs.notifyObservers(null, "activity-streams-places-cache-complete", null);
       }
+
+      // Call myself when cache expires to repopulate.
+      // This is needed because some of the queries are time dependent (for example,
+      // highlights excludes links from the past 30 minutes).
+      if (this._placesCacheTimeoutID) {
+        clearTimeout(this._placesCacheTimeoutID);
+      }
+      this._placesCacheTimeoutID = setTimeout(() => {
+        this._asyncBuildPlacesCache();
+      }, this.options.placesCacheTimeout);
     }
   }),
 
@@ -471,6 +482,7 @@ ActivityStreams.prototype = {
   unload(reason) { // eslint-disable-line no-unused-vars
     let defaultUnload = () => {
       clearTimeout(this._previewCacheTimeoutID);
+      clearTimeout(this._placesCacheTimeoutID);
       this._previewProvider.uninit();
       NewTabURL.reset();
       this.workers.clear();
