+++ test/test-ActivityStreams-async-preview-cache-rebuild.js
@@ -1,4 +1,4 @@
-/* globals XPCOMUtils, Services */
+/* globals XPCOMUtils */
 
 "use strict";
 const {Cu} = require("chrome");
@@ -8,9 +8,9 @@ const simplePrefs = require("sdk/simple-prefs");
 const {ActivityStreams} = require("lib/ActivityStreams");
 const {PlacesProvider} = require("lib/PlacesProvider");
 const {PreviewProvider} = require("lib/PreviewProvider");
+const {makeCachePromise, makeCountingCachePromise} = require("./lib/cachePromises");
 
 Cu.import("resource://gre/modules/XPCOMUtils.jsm");
-Cu.import("resource://gre/modules/Services.jsm");
 
 XPCOMUtils.defineLazyModuleGetter(this, "PlacesUtils",
                                   "resource://gre/modules/PlacesUtils.jsm");
@@ -23,36 +23,6 @@ XPCOMUtils.defineLazyModuleGetter(this, "Task",
 
 let gInitialCachePref = simplePrefs.prefs["query.cache"];
 let gPreviewProvider;
-
-let makeCachePromise = (name) => {
-  return new Promise(resolve => {
-    let precacheNotif = `activity-streams-${name}-cache-complete`;
-    let waitForCache = (subject, topic, data) => {
-      if (topic === precacheNotif) {
-        Services.obs.removeObserver(waitForCache, precacheNotif);
-        resolve();
-      }
-    };
-    Services.obs.addObserver(waitForCache, precacheNotif);
-  });
-};
-
-let makeCountingCachePromise = (name, target) => {
-  return new Promise(resolve => {
-    let count = 0;
-    let precacheNotif = `activity-streams-${name}-cache-complete`;
-    let waitForCache = (subject, topic, data) => {
-      if (topic === precacheNotif) {
-        count++;
-        if (count === target) {
-          Services.obs.removeObserver(waitForCache, precacheNotif);
-          resolve(count);
-        }
-      }
-    };
-    Services.obs.addObserver(waitForCache, precacheNotif);
-  });
-};
 
 exports["test preview cache repopulation works"] = function*(assert) {
   let placesCachePromise;
@@ -69,6 +39,19 @@ exports["test preview cache repopulation works"] = function*(assert) {
   let numRepopulations = yield previewsCountPromise;
 
   assert.equal(numRepopulations, expectedRepopulations, "preview cache successfully repopulated periodically");
+  app.unload();
+};
+
+exports["test places cache repopulation works"] = function*(assert) {
+  let placesCachePromise = makeCachePromise("places");
+  let app = new ActivityStreams({placesCacheTimeout: 100});
+  yield placesCachePromise;
+
+  let expectedRepopulations = 3;
+  let placesCountPromise = makeCountingCachePromise("places", expectedRepopulations);
+  let numRepopulations = yield placesCountPromise;
+
+  assert.equal(numRepopulations, expectedRepopulations, "places cache successfully repopulated periodically");
   app.unload();
 };
